Starting a branch to add more information into the annotate web page.
I added a link <strong>p</strong> to the annotation of the version previous to
that line change, and a link <strong>d</strong> to the checkin file diff that
introduced the given line.

I'd want also the links to bring the users to the concerning line of annotate; we should
have something like line numbering and line anchors, like the artifact page.

FossilOrigin-Name: 9e3b6ad052374381da5b656710701b824db0e861