I think I fix the previous commit, that should avoid committing a ticket manifest if it has no changes. It was ignoring the ticket appends.
In the case of only appending, the user was not notified not properly redirected, and the ticket
chagnes were not committed.

Still, there should be some kind of warning to the user, if it ends up in such situation, I think.

FossilOrigin-Name: f02946aa8b4e27124c1e5840ef907a60dc875025