Begin work merging new_brlist_page() into brlist_page().
Annoyance: The new open/closed checkboxes don't quite work the way I want
them to work.  Deselecting both has the effect of selecting both.  This makes
sense in that if neither the open nor closed query parameters are given,
everything should be displayed, rather than nothing.  Yet, the user interface
would be nicer if deselecting the only selected checkbox causes the other
checkbox to be selected.  This can't be done with straight HTML because
deselecting the checkbox removes it from the query string, and the server does
not know which checkbox was deselected so it cannot force the other one to
become selected.  As far as I know, JavaScript is required.

Bug: Non-aggregate HAVING clause expressions are evaluated on an arbitrarily
chosen row within each group.  This breaks closed and open filtering for the
command-line branch command and the new checkboxes.  But shouldn't this also
break non-aggregate result expressions?

Wish: Filter on normal/hidden branches too.

Wish: If a merged-to branch is hidden, don't show it in the resolution column,
instead show something else.  The latest merged-to?  The earliest merged-to?

FossilOrigin-Name: a6a4482056851df4b3d194d1b524d6e252e92df0