Fossil's chroot feature drops root permissions based on file ownership, but since the container was built with everything-root, its HTTP hit handling children would run as whatever host-side UID/GID pair you used for file ownership.  What happened next was complex.
If you let the container create the repo internally, it would be owned
as root, so it would drop root permissions for…root!  This isn't
super-bad, since Fossil is presumed secure and is double-jailed besides.
The risk is, if anyone works out an RCE for Fossil, they might be able
to get it to create raw sockets or do various other types of escapes
despite the double-jail dance.

Attaching a Docker volume brings external permisssions into the
container.  We were recommending a "chown 0" command on the shared
volume to make it similar to the in-container case, but that opens you
to the same risks above.  If you ignored this and used host-side UID/GID
pairs, Fossil would then be left running under IDs that didn't exist
internally, which could cause assorted weirdness.

We're now creating an explicit "fossil" user/group pair inside the
container and recommending that Docker volumes use these IDs for
copied-in files to batten down something that shouldn't've been left
flapping.

Updated build.wiki to cover all this.

FossilOrigin-Name: ba21bc0b8fcf754d20de4a65a5640c9f7de441eac5bcbfa9c892a3f202cf48e0