Added three new block-element classes to the skin to replace use of "blockquote" with semantic alternatives:
  *  "indent" for cases where the only purpose was to indent things
  *  "blockquote" where you want an actual block quote in a wiki doc and
     thus can't use a "blockquote" element due to the historical
     misuse of this tag, preventing us from doing the same for it as we
     do for MD
  *  "sidebar" when the reason you wanted an indented paragraph was to
     create a sidebar; this is styled much like a blockquote in MD (or
     the new "blockquote" class in wiki docs) but floats to the right.

Using all three of these in the remaining www/ docs that still used
explicit "blockquote" tags.

(If you're wondering why the prior commit claims to have dealt with all
of these tags when this commit belies that, it's because this work was
done en masse, then split up into two related commits to keep the commit
messages more on-point.  It was easier to commit these two chunks of
work in this order. It would have required a complicated dance with the
stash to commit them in the more logical order.)

FossilOrigin-Name: 56c7c2cb1393301f5a26e5e3019176039227dedf6a8a8669760bc4aff9f8fb3b