Reworked the handling of the three "selected" row cases in /timeline (current, first-in-range and last-in-range) so the borders and highlighting extend to the left margin.  The "current" case (seen under "fossil ui" only, to indicate the ckout version) wasn't a problem, but the other two use background colors potentially confusable with branch colors.  This makes the timelines more like that of the Fossil 2.23 and prior default skin (now called Étienne) but with rounded corners in the modern modes, matching the rest of the timeline.
This occurs in all timeline modes, including the classic, compact, and
verbose views.  The result is that the corners for these selected row
types are rounded, but because they tend to be compact, it gives a
full-round look on the ends.  Avoiding this would require considerably
more verbose CSS, and it's plenty intricate as it is.

FossilOrigin-Name: 9568091a9beee72fae8fd189f09bcf6da7cb3f1fc71de3129b03ddb4bce3fd55